pyproject.toml
[tool.poetry]
name = "rentaxon-backend"
version = "0.1.0"
description = ""
authors = ["moayadalhaj <moayad.alhaj21@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.0"
asgiref = "^3.4.1"
django-cors-headers = "^3.10.1"
django-environ = "^0.8.1"
djangorestframework-simplejwt = "^5.0.0"
djangorestframework = "^3.12.4"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.2"
PyJWT = "^2.3.0"
pytz = "^2021.3"
sqlparse = "^0.4.2"
whitenoise = "^5.3.0"
django-phonenumber-field = "^6.0.0"
phonenumbers = "^8.12.39"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
